requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
requests==2.31.0

# Computer Vision and Face Recognition
opencv-python==4.8.1.78
deepface==0.0.79
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.1

# Additional dependencies for DeepFace
mtcnn==0.1.1
retina-face==0.0.13
facenet-pytorch==2.5.3

# Data handling
pandas==2.0.3
scikit-learn==1.3.0

# Optional: For better performance
python-multipart==0.0.6
